+++ src/aws/ec2-wrapper.ts
@@ -1,11 +1,29 @@
 import { EC2 } from "aws-sdk";
 import { getAWSPagedResults } from "../utils";
+import { VPC, VPCDiscovery } from "../types";
 
-class EC2Wrapper {
+export class EC2Wrapper {
   public ec2: EC2
 
   constructor (credentials: any) {
     this.ec2 = new EC2(credentials);
+  }
+
+  /**
+   * Gets the desired vpc with the designated subnets and security groups
+   * that were set in serverless config file
+   * @returns {Promise<object>}
+   */
+  public async getVpcConfig (vpcDiscovery: VPCDiscovery): Promise<VPC> {
+    const vpc: VPC = {};
+    const vpcId = await this.getVpcId(vpcDiscovery.vpcName);
+    if (vpcDiscovery.subnetNames) {
+      vpc.subnetIds = await this.getSubnetIds(vpcId, vpcDiscovery.subnetNames);
+    }
+    if (vpcDiscovery.securityGroupNames) {
+      vpc.securityGroupIds = await this.getSecurityGroupIds(vpcId, vpcDiscovery.securityGroupNames);
+    }
+    return vpc;
   }
 
   /**
@@ -114,7 +132,7 @@ class EC2Wrapper {
     }
 
     const missingGroupsNames = securityGroupNames.filter((groupName) => {
-      // collect subnets by name
+      // collect security groups by name
       const securityGroupsByName = securityGroups.filter((securityGroup) => {
         return securityGroup.GroupName === groupName;
       });
@@ -131,5 +149,3 @@ class EC2Wrapper {
     return securityGroups.map((group) => group.GroupId);
   }
 }
-
-export = EC2Wrapper