+++ test/integration-tests/base.ts
@@ -1,9 +1,9 @@
-import randomstring from "randomstring"
+import randomstring from "randomstring";
 
-const TEST_VPC_NAME = process.env.TEST_VPC_NAME
+const TEST_VPC_NAME = process.env.TEST_VPC_NAME;
 
 if (!TEST_VPC_NAME) {
-  throw new Error("TEST_VPC_NAME environment variable not set")
+  throw new Error("TEST_VPC_NAME environment variable not set");
 }
 
 function getRandomString (): string {
@@ -11,10 +11,10 @@ function getRandomString (): string {
     capitalization: "lowercase",
     charset: "alphanumeric",
     length: 5
-  })
+  });
 }
 
 export {
   getRandomString,
   TEST_VPC_NAME
-}
+};